final printf

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: dsayumi- <dsayumi-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/12 19:56:19 by dsayumi-          #+#    #+#             */
-/*   Updated: 2023/08/17 22:29:19 by dsayumi-         ###   ########.fr       */
+/*   Updated: 2023/08/18 21:42:09 by dsayumi-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@ int		ft_putnbr(int n);
 int		ft_strlen(const char *s);
 char	ft_putptr(unsigned long nbr, int i);
 int		ft_putunsign(unsigned int nbr);
-int		ft_puthex_upper(unsigned long n);
-int		ft_puthex_low(unsigned long n);
+int		ft_puthex_upper(unsigned int n);
+int		ft_puthex_low(unsigned int n);
 int		ft_putstr(char *s);
 int		convert_params(char *str, int i, va_list vl);
